+++ integration-tests/test/kotlin/com/javiersc/integration/tests/gradle/extensions/DelegatedPropertiesTest.kt
@@ -16,7 +16,8 @@ class DelegatedPropertiesTest : GradleTestKitTest() {
                         |DELEGATED 1
                         |DELEGATED 2
                     """
-                        .trimMargin())
+                        .trimMargin()
+                )
         }
     }
 }