+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 ### Updated
 
-- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.7.6`
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.7.7`
 - `gradle -> 8.10`
 
 ## [1.0.0-alpha.34] - 2024-02-03